added relation between user and post models

+++ rest-api/controllers/feed.js
@@ -5,32 +5,30 @@ const path = require('path');
 const isAuth = require('../middleware/is-auth');
 
 const Post = require('../models/post');
+const User = require('../models/user');
 
 
-exports.getPosts = (req, res, next) => {
+exports.getPosts = async (req, res, next) => {
     const currentPage = req.query.page || 1;
     const perPage = 2;
-    let totalItems;
-    Post.find().countDocuments()
-    .then(count=> {
-        totalItems = count;
-        return Post.find().skip((currentPage - 1) * perPage).limit(perPage);
+    try {
+    const totalItems = await Post.find().countDocuments()
+    const posts = await Post.find().populate('creator').skip((currentPage - 1) * perPage).limit(perPage);
+
+    res.status(200).json({
+        message: 'Fetched posts successfully',
+        posts,
+        totalItems,
     })
-    .then(posts=> {
-        res.status(200).json({
-            message: 'Fetched posts successfully',
-            posts,
-            totalItems,
-        })
-    }).catch(err=>{
+    } catch(err){
         if(!err.statusCode){
             err.statusCode = 500;
         }
         next(err);
-    })
+    }
 }
 
-exports.createPost = (req, res, next) => {
+exports.createPost =async (req, res, next) => {
     const errors = validationResult(req);
     if (!errors.isEmpty()) {
         const error = new Error('Validation failed, entered data is incorrect');
@@ -45,46 +43,56 @@ exports.createPost = (req, res, next) => {
 
     const imageUrl = req.file.path.replace("\\" ,"/");
 
-
+    let creator;
     const post = new Post({
         title: req.body.title,
         content: req.body.content,
         imageUrl: imageUrl,
-        creator: {name: 'Maximilian'},
+        creator: req.userId,
     })
-    post.save().then(result => {
-        console.log(result);
+    try {
+        await post.save();
+
+        const user = await User.findById(req.userId);
+        creator = user;
+        user.posts.push(post)
+
+        await user.save();
+
         res.status(201).json({
             message: 'Post created successfully!',
-            post: result,
+            post: post,
+            creator: {_id: creator._id, name: creator.name}
         })
-    }).catch(err => {
-        if (!err.statusCode) {
+    } catch(err){
+        if(!err.statusCode){
             err.statusCode = 500;
         }
-        next(err)
-    })
+        next(err);
+    }
 }
 
 
     //Create post in db
 
-    exports.getPost = (req, res, next) => {
+    exports.getPost = async (req, res, next) => {
         const postId = req.params.postId;
-        Post.findById(postId)
-            .then(post=>{
-                if(!post){
-                    const error = new Error('Could not find post');
-                    error.statusCode = 404;
-                    throw error;
-                }
-                res.status(200).json({message: 'Post fetched', post: post})
-            }).catch(err=>{
+        try {
+            const post = await Post.findById(postId)
+            if (!post) {
+                const error = new Error('Could not find post');
+                error.statusCode = 404;
+                throw error;
+            }
+
+            res.status(200).json({message: 'Post fetched', post: post})
+        } catch(err){
             if(!err.statusCode){
                 err.statusCode = 500;
             }
             next(err);
-        })
+        }
+
     }
 
     exports.updatePost = (req, res, next ) => {
@@ -107,14 +115,23 @@ exports.createPost = (req, res, next) => {
             throw error;
         }
         Post.findById(postId).then(post=>{
+
             if(!post){
                 const error = new Error('Could not find post');
                 error.statusCode = 404;
                 throw error;
             }
+
+            if(post.creator.toString() !== req.userId) {
+                const error = new Error('Not authorized');
+                error.statusCode = 403;
+                throw error;
+            }
+
             if(imageUrl !== post.imageUrl) {
                 clearImage(post.imageUrl)
             }
+
             post.title = title;
             post.imageUrl = imageUrl;
             post.content = content;
@@ -138,9 +155,21 @@ exports.createPost = (req, res, next) => {
                     error.statusCode = 404;
                     throw error;
                 }
+
+                if(post.creator.toString() !== req.userId) {
+                    const error = new Error('Not authorized');
+                    error.statusCode = 403;
+                    throw error;
+                }
+
                 clearImage(post.imageUrl);
                 return Post.findByIdAndDelete(postId);
 
+        }).then(result=> {
+            return User.findById(req.userId);
+        }).then(user=> {
+            user.posts.pull(postId);
+            return user.save();
         }).then(result=>{
             res.status(200).json({message: 'Delete post'});
         }).catch(err=>{
